refactor: separate business logic from ProjectDetailView
